obs_make_hysplit_nc4

Usage

obs_make_hysplit_nc4(
  ident,
  ncname = paste("hysplit", ident, ".nc", sep = ""),
  origident = FALSE,
  part = FALSE,
  foot = FALSE,
  horizconv = FALSE,
  avghorizconv = FALSE,
  dvmr = FALSE,
  dvmrmean = FALSE,
  dvmrstd = FALSE,
  runinfo = FALSE,
  check = FALSE,
  emit = FALSE,
  init = FALSE,
  avginit = FALSE,
  partdatname = "part",
  footname = NULL,
  horizconvname = NULL,
  avghorizconvname = NULL,
  dvmrname = NULL,
  dvmrmeanname = NULL,
  dvmrstdname = NULL,
  initname = NULL,
  avginitname = NULL,
  horizconvdatunits = NULL,
  avghorizconvdatunits = NULL,
  dvmrdatunits = NULL,
  dvmrmeandatunits = NULL,
  dvmrstddatunits = NULL,
  initdatunits = NULL,
  avginitdatunits = NULL,
  origidentdat = NULL,
  origutctime.format = NULL,
  partdat = NULL,
  footdat = NULL,
  horizconvdat = NULL,
  avghorizconvdat = NULL,
  dvmrdat = NULL,
  dvmrmeandat = NULL,
  dvmrstddat = NULL,
  runinfodat = NULL,
  checkdat = NULL,
  emitdat = NULL,
  initdat = NULL,
  avginitdat = NULL,
  partpath = NULL,
  checkpath = partpath,
  footpath = paste(partpath, "footprints/", sep = ""),
  horizconvpath = paste(partpath, "convolve/", sep = ""),
  avghorizconvpath = paste(partpath, "convolve/", sep = ""),
  initpath = paste(partpath, "boundary/", sep = ""),
  avginitpath = paste(partpath, "boundary/", sep = ""),
  runinfopath = NULL,
  targetdir,
  errorlog = paste("ncerrors", ident, ".txt", sep = ""),
  appendnc = FALSE,
  plotfoot = FALSE,
  plotmeantrajec = FALSE,
  sourcepath = NULL,
  charmissval = NULL,
  global.att.list = NULL,
  global.att.files = NULL,
  ident.start = NULL
)

Arguments

ident

identifier

ncname

netcdf name file format hysplitid.nc

origident

logical, default FALSE, include most accurate receptor agl,lat,lon, utctime fields, before rounding

part

logical, default FALSE, include particle array

foot

logical, default FALSE, include foot

horizconv

logical, default FALSE, include results from horizontally convolving footprints with fluxes for each particle and summing particle over time

avghorizconv

logical, default FALSE, include results from horizontally convolving footprints with fluxes for each particle and summing particle over time, avg particles

dvmr

logical, default FALSE, include delta volume mixing ratio results from convolve-fast.r (replaces horizconv and avghorizconv functionality)

dvmrmean

logical, default FALSE, include climatological mean of delta volume mixing ratio results

dvmrstd

logical, default FALSE, include climatological standard dev of volume mixing ratio results

runinfo

logical, default FALSE, text summary of run parameters

check

logical, default FALSE, output from Trajeccheck

emit

logical, default FALSE, emission window description

init

logical, default FALSE, include init mixing ratio values

avginit

logical, default FALSE, include avginit mixing ratio value

partdatname

string, default 'part', name of particle table, with the following convention: part => whole particle table (default) partfoot => particle table thinned to retain rows where foot > 0 part3d => particle table thinned to retain rows every nhrs endpts => particle table thinned to retain rows with trajectory endpts

footname

default NULL, name of footprint to distinguish among multiple footprints -e.g. foot1, foot2, etc

horizconvname

name of horizontal convolve mixing ratio values (should reflect species nd source - e.g. co2horizconvCT, co2horizconvvul, etc)

avghorizconvname

name of avg horizontal convolve mixing ratio (should reflect species and source - e.g. co2avghorizconvCT, co2avghorizconvvul, etc)

dvmrname

name of delta volume mixing ratio results (should reflect species and source - e.g. co2dvmrCT, co2dvmrvul, etc)

dvmrmeanname

name of climatological mean of delta volume mixing ratio results (should reflect species and source - e.g. co2dvmrmeanCT, co2dvmrmeanvul, etc)

dvmrstdname

name of climatological std of delta volume mixing ratio results (should reflect species and source - e.g. co2dvmrstdCT, co2dvmrstdvul, etc)

initname

name of init mixing ratio values (should reflect species and source) (e.g. co2initCT, co2initGV, coinitGV, etc)

avginitname

name of average init mixing ratio (should reflect species and source) (e.g. co2avginitCT, co2avginitGV, coavginitGV, etc)

horizconvdatunits

default NULL, string indicating assigned units to horizconvdat

avghorizconvdatunits

default NULL, string indicating assigned units to avghorizconvdat

dvmrdatunits

default NULL, string indicating assigned units to dvmrdat

dvmrmeandatunits

default NULL, string indicating assigned units to dvmrmeandat

dvmrstddatunits

default NULL, string indicating assigned units to dvmrstddat

initdatunits

default NULL, string indicating assigned units to initdat

avginitdatunits

default NULL, string indicating assigned units to avginitdat

origidentdat

default NULL, list of original receptor agl, lat, lon, utctime fields

origutctime.format

default NULL, string format of origutctime from receptor list e.g. '

partdatdefault NULL, passed in R object (use getr by default)

footdatdefault NULL, passed in R object (use getr by default)

horizconvdatdefault NULL, passed in R object (use getr by default)

avghorizconvdatdefault NULL, passed in R object (use getr by default)

dvmrdatdefault NULL, matrix of dvmr results

dvmrmeandatdefault NULL, matrix of dvmrmean results

dvmrstddatdefault NULL, matrix of dvmrstd results

runinfodatdefault NULL, passed in R object (use getr by default)

checkdatdefault NULL, passed in R object (use getr by default)

emitdatdefault NULL, data frame/2D array

initdatdefault NULL, passed in R object (use getr by default)

avginitdatdefault NULL, passed in R object (use getr by default)

partpathdefault NULL, location of particle file

checkpathdefault partpath, location of particle file

footpathpaste(partpath, "footprints/", sep=""), location of footprint file

horizconvpathpaste(partpath, "convolve/", sep="") location of horizontal convolve summarized results file

avghorizconvpathpaste(partpath, "convolve/", sep="") location of horizontal convolve summarized results file

initpathpaste(partpath, "boundary/", sep="") location of init file

avginitpathpaste(partpath, "boundary/", sep="") location of init file

runinfopathdefault NULL, location of runinfo file

targetdirdestination directory

errorlogpaste("ncerrors",ident,".txt",sep=""), name of errorlog file; use NA to suppress

appendncdefault FALSE, to append to nc

plotfootdefault FALSE, to plot footprints

plotmeantrajecdefault FALSE, to plot mean trajectory

sourcepathdefault NULL

charmissvaldefault NULL, missing value for character variables (ncvar_def requires NA or NULL depending on the version of ncdf4)

global.att.listdefault NULL, list of global attributes (only used if !appendnc)

global.att.filesdefault NULL, list of filenames whose contents are turned into global attributes (only used if !appendnc and global.att.list is empty)

ident.startdefault NULL, time related to hysplit start time (may different from nearest hour to ident if partperhymodelc was turned on)

return dates makes nc Convention is that the Julian date of the epoch is 1.0 (not 0!) Other helpers legacy: obs_decimal_to_POSIX(), obs_id2pos(), obs_info2id(), obs_interpret_udunits_time(), obs_julian(), obs_load_ncdf(), obs_traj_foot() helpers legacy