obs_traj_foot

return trajectory

Usage

obs_traj_foot(
  ident,
  part = NULL,
  timelabel = NULL,
  pathname = "",
  foottimes = c(0, 360),
  zlim = c(0, 0),
  coarse = 1,
  dmassTF = TRUE,
  numpix.x = 376,
  numpix.y = 324,
  lon.ll = -145,
  lat.ll = 11,
  lon.res = 1/4,
  lat.res = 1/6,
  read.r = TRUE,
  read.nc = TRUE,
  nparstilt = NULL,
  eps.global = 0.1
)

Arguments

ident

character value specifying the trajectory ensemble to look at

part

object containing particle run; if NULL, then determine from ident and pathname

pathname

path where object with particle locations is saved

foottimes

vector of times between which footprint or influence will be integrated

zlim

(if not default 0,0): vertical interval for which particle distribution is looked at

coarse

degrade resolution (for aggregation error): 0: only 20 km resolution; 1-16: dynamic resolution, but limited highest resolution coarse: (1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16) coarsex:c(1,1,2,2,2,4,4,4,8, 8, 8,16,16,16,32,32) #factors by which grids have been made coarser coarsey:c(1,2,1,2,4,2,4,8,4, 8,16, 8,16,32,16,32)#e.g., '4' means grid is 4 times coarser

dmassTF

default TRUE weighting by accumulated mass due to violation of mass conservation in met fields

numpix.x

default 376, number of pixels in x directions in grid

numpix.y

default 324, number of pixels in y directions in grid

lon.ll

default -145, lower left corner of grid (longitude of southwest corner of southwest corner gridcell)

lat.ll

default 11, lower left corner of grid (latitude of southwest corner of southwest corner gridcell)

read.r

default TRUE

read.nc

default TRUE

nparstilt

default NULL, number of particles stilt was run with (required in order to account for those cases where a thinned particle table that may not contain all particle indices is used)

eps.global

default 0.1

Value

return footprint

See also

Examples